requirements.txt

flask
nltk
pyyaml
matplotlib
numpy
pandas
